requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Limiter==3.7.0
email-validator==2.1.1
python-dotenv==1.0.1
psycopg[binary]==3.2.13
gunicorn==22.0.0
cloudinary==1.41.0
markdown==3.6
bleach==6.1.0
reportlab==4.1.0
pywebpush==1.14.0
requests==2.31.0
urllib3==2.2.3
charset-normalizer==3.3.2
openai==1.51.0
httpx>=0.27.0,<0.28.0
celery[redis]==5.5.3
websockets==12.0
pytest==8.1.1
pytest-cov==5.0.0
